backend/requirements.txt
# Core
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
python-dotenv>=1.0.1
pyyaml>=6.0.1

# LLM routing
litellm>=1.40.0

# Coding engine
aider-chat>=0.53.0

# GitHub + Git
PyGithub>=2.3.0
GitPython>=3.1.43

# Scheduling
apscheduler>=3.10.4

# SQLite timezone support
pytz>=2024.1

# Linting (for linting the target repo — optional, swap as needed)
ruff>=0.4.0
